requirements.txt
absl-py==0.4.1
actionlib==1.11.13
angles==1.9.11
astor==0.7.1
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
bondpy==1.8.1
camera-calibration==1.12.22
camera-calibration-parsers==1.11.13
catkin==0.7.8
catkin-pkg==0.4.8
configparser==3.5.0
controller-manager-msgs==0.13.0
cv-bridge==1.12.8
cycler==0.10.0
decorator==4.3.0
diagnostic-analysis==1.9.2
diagnostic-common-diagnostics==1.9.2
diagnostic-updater==1.9.2
docutils==0.14
dynamic-reconfigure==1.5.49
entrypoints==0.2.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
gast==0.2.0
gazebo-plugins==2.5.14
gazebo-ros==2.5.14
gencpp==0.5.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.9
gennodejs==2.0.1
genpy==0.6.7
grpcio==1.14.2
h5py==2.8.0
html5lib==0.9999999
image-geometry==1.12.7
interactive-markers==1.11.3
ipaddress==1.0.22
ipykernel==4.8.2
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
laser-geometry==1.6.4
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
message-filters==1.12.12
mistune==0.8.3
mock==2.0.0
moveit-commander==0.9.11
moveit-ros-planning-interface==0.9.11
moveit-ros-visualization==0.9.11
mpmath==1.0.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
np-utils==0.5.5.0
numpy==1.14.5
pandocfilters==1.4.2
pathlib2==2.3.2
pbr==4.2.0
pexpect==4.6.0
pickleshare==0.7.4
pluginlib==1.11.2
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-qt-binding==0.3.3
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qt-dotgraph==0.3.8
qt-gui==0.3.8
qt-gui-cpp==0.3.8
qt-gui-py-common==0.3.8
qtconsole==4.4.1
resource-retriever==1.12.3
rosbag==1.12.12
rosboost-cfg==1.13.6
rosclean==1.13.6
roscreate==1.13.6
rosgraph==1.12.12
roslaunch==1.12.12
roslib==1.13.6
roslint==0.11.0
roslz4==1.12.12
rosmake==1.13.6
rosmaster==1.12.12
rosmsg==1.12.12
rosnode==1.12.12
rosparam==1.12.12
rospkg==1.1.7
rospy==1.12.12
rospy-message-converter==0.4.0
rosservice==1.12.12
rostest==1.12.12
rostopic==1.12.12
rosunit==1.13.6
roswtf==1.12.12
rqt-action==0.4.9
rqt-bag==0.4.11
rqt-bag-plugins==0.4.11
rqt-console==0.4.8
rqt-dep==0.4.8
rqt-graph==0.4.9
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.11
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.8
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.7
rqt-robot-steering==0.5.8
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.7
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.5.8
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rviz==1.12.16
scandir==1.9.0
scikit-learn==0.20.0
scipy==1.1.0
Send2Trash==1.5.0
sensor-msgs==1.12.5
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.1
srdfdom==0.4.2
subprocess32==3.5.2
sympy==1.2
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.17
tf2-kdl==0.5.17
tf2-py==0.5.17
tf2-ros==0.5.18
topic-tools==1.12.12
tornado==5.1
traitlets==4.3.2
urdfdom-py==0.3.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.0
wiimote==1.11.0
xacro==1.11.3
